can add decisions

+++ Decision System/Core/Models/GroupDecision.cs
@@ -9,10 +9,10 @@ namespace Core.Models
     public class GroupDecision
     {
         public Group Group { get; set; }
-        public int GroupId { get; set; }
+        public long GroupId { get; set; }
 
         public Decision Decision { get; set; }
-        public int DecisionId { get; set; }
+        public long DecisionId { get; set; }
 
         public GroupDecision()
         {
